Extract shared Spring Data support
