+++ kuery-client-spring-data-common/src/main/kotlin/dev/hsbrysk/kuery/spring/data/internal/ValueClasses.kt
@@ -1,11 +1,9 @@
-package dev.hsbrysk.kuery.spring.jdbc.internal
+package dev.hsbrysk.kuery.spring.data.internal
 
+import dev.hsbrysk.kuery.core.KueryClientInternalApi
 import org.springframework.util.ClassUtils
 import java.lang.reflect.Method
 
-// NOTE: This object is intentionally duplicated in the r2dbc module
-// (dev.hsbrysk.kuery.spring.r2dbc.internal.ValueClasses); there is no shared Spring-dependent
-// module to host it. Keep both copies in sync.
 /**
  * Reflection over the compiler-generated JVM ABI of Kotlin value classes (`@JvmInline`). This is
  * the single owner of the mangled member names — `unbox-impl` (used on the write path to bind a
@@ -13,7 +11,8 @@ import java.lang.reflect.Method
  * read path to box an underlying value back into a value class). Plain Java reflection keeps the
  * per-parameter bind hot path free of kotlin-reflect machinery.
  */
-internal object ValueClasses {
+@KueryClientInternalApi
+public object ValueClasses {
     // ClassValue (instead of a Class-keyed map) so cached entries are tied to the class itself
     // and never pin user classloaders (e.g. across Spring DevTools restarts).
     private val unboxMethods = object : ClassValue<Method>() {
@@ -28,13 +27,13 @@ internal object ValueClasses {
      * Returns the underlying value of a boxed value class instance. May be null when the
      * underlying type is nullable (e.g. `value class Opt(val value: String?)`).
      */
-    fun unbox(value: Any): Any? = unboxMethods.get(value.javaClass).invoke(value)
+    public fun unbox(value: Any): Any? = unboxMethods.get(value.javaClass).invoke(value)
 
     /**
      * The JVM type of the underlying value, e.g. `UserName` -> `String`. Primitives are boxed
      * (`Int` -> `java.lang.Integer`) because callers use this as an object array component type.
      */
-    fun underlyingType(clazz: Class<*>): Class<*> =
+    public fun underlyingType(clazz: Class<*>): Class<*> =
         ClassUtils.resolvePrimitiveIfNecessary(unboxMethods.get(clazz).returnType)
 
     /**
@@ -43,7 +42,7 @@ internal object ValueClasses {
      * does not expose exactly one of each (e.g. a value class with a secondary constructor), in
      * which case callers fall back to the kotlin-reflect constructor.
      */
-    fun fastBoxOrNull(valueClass: Class<*>): FastBox? {
+    internal fun fastBoxOrNull(valueClass: Class<*>): FastBox? {
         val constructorImpl = valueClass.singleDeclaredMethodOrNull("constructor-impl") ?: return null
         val boxImpl = valueClass.singleDeclaredMethodOrNull("box-impl") ?: return null
         return FastBox(constructorImpl, boxImpl)
@@ -53,7 +52,7 @@ internal object ValueClasses {
         declaredMethods.singleOrNull { it.name == name }?.apply { isAccessible = true }
 
     /** Boxes an underlying value into a value class via its static `constructor-impl`/`box-impl`. */
-    class FastBox internal constructor(
+    internal class FastBox(
         private val constructorImpl: Method,
         private val boxImpl: Method,
     ) {
