+++ settings.gradle.kts
@@ -31,5 +31,6 @@ include("kuery-client-compiler:functional-test-auto-trim")
 include("kuery-client-core")
 include("kuery-client-gradle-plugin")
 include("kuery-client-gradle-plugin:functional-test")
+include("kuery-client-spring-data-common")
 include("kuery-client-spring-data-jdbc")
 include("kuery-client-spring-data-r2dbc")
